Added Core Features
Added a new attribute and some functionality to the Service Manager as well as fixed some typos.

Attribute:
[DoNotInitializeOnLoad] can now be placed on singletons you wish to initialize manually instead of after assemblies have been loaded!

Methods:
Added 2 new methods
GetOrStartService<T>() && ForceStartService<T>()

these methods will allow you to either fetch and existing service or create a new one, or if you need to force start a service whether it exists or not, terminating an active instance if one exists!