Refactoring and Extension
Added a few much needed functions to allow users to add their own services to the service manager when using the [DoNotInitializeOnLoad] attribute, allowing them to create services which can be drag and dropped into the scene.
Refactored some of the namespaces and changes a few things around to make it more intuitive in larger workflows.